+++ crates/filter-list-manager/src/io/fetch_by_schemes.rs
@@ -1,5 +1,5 @@
 use crate::filters::parser::parser_error::FilterParserError;
-use crate::io::{http::HttpClient, read_filter_file, url_schemes::UrlSchemes};
+use crate::io::{http::HttpClient, read_file_by_url, url_schemes::UrlSchemes};
 
 /// Synchronously fetch contents from absolute url
 pub(crate) fn fetch_by_scheme(
@@ -8,7 +8,7 @@ pub(crate) fn fetch_by_scheme(
     request_timeout: i32,
 ) -> Result<String, FilterParserError> {
     match scheme {
-        UrlSchemes::File => read_filter_file(absolute_url).map_err(Into::into),
+        UrlSchemes::File => read_file_by_url(absolute_url).map_err(Into::into),
         UrlSchemes::Https | UrlSchemes::Http => {
             HttpClient::sync_get_file_contents(absolute_url, request_timeout)
                 .map_err(FilterParserError::Network)