+++ crates/filter-list-manager/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.11] - 2025-01-17
+
+### Fixed
+- Unnecessary filter rules selection in `save_disabled_rules`
+- `file:///` urls support in `IndexesParser`
+
+[1.1.11]: https://github.com/AdguardTeam/FilterListManager/compare/flm-1.1.10...flm-1.1.11
+
 ## [1.1.10] - 2024-12-19
 
 ### Fixed